# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Kingsway Backend API
  version: 0.2.0
  description: Kingsway Operation Center backend HTTP contract. Authenticated JSON POST write endpoints accept optional Idempotency-Key for safe same-payload retry.
servers:
  - url: http://127.0.0.1:8080
security:
  - bearerAuth: []
...
paths:
...
  /v1/students:
    get:
      tags: [academic]
      parameters:
        - $ref: "#/components/parameters/BranchID"
        - name: status
          in: query
          schema:
            enum: [active, left, graduated]
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          $ref: "#/components/responses/StudentList"
    post:
      tags: [academic]
      requestBody:
        $ref: "#/components/requestBodies/CreateStudent"
      responses:
        "201":
          $ref: "#/components/responses/Student"
...
  /v1/students/{student_id}:
    get:
      tags: [academic]
      parameters:
        - $ref: "#/components/parameters/StudentID"
      responses:
        "200":
          $ref: "#/components/responses/Student"
    patch:
      tags: [academic]
      parameters:
        - $ref: "#/components/parameters/StudentID"
      requestBody:
        $ref: "#/components/requestBodies/UpdateStudent"
      responses:
        "200":
          $ref: "#/components/responses/Student"
...
  /v1/classes/{class_id}/students:
    get:
      tags: [academic]
      parameters:
        - $ref: "#/components/parameters/ClassID"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Class enrollments.
    post:
      tags: [academic]
      parameters:
        - $ref: "#/components/parameters/ClassID"
      requestBody:
        $ref: "#/components/requestBodies/EnrollStudent"
      responses:
        "201":
          description: Created enrollment.
...
  /v1/files:
    get:
      tags: [files]
      parameters:
        - $ref: "#/components/parameters/BranchID"
        - name: owner_type
          in: query
          schema:
            type: string
        - name: owner_id
          in: query
          schema:
            type: string
        - name: purpose
          in: query
          schema:
            enum: [material, payment_receipt, event, passport, diploma, profile_photo, exam_writing]
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          $ref: "#/components/responses/FileList"
    post:
      tags: [files]
      requestBody:
        $ref: "#/components/requestBodies/RegisterFile"
      responses:
        "201":
          $ref: "#/components/responses/File"
...
  /v1/files/{file_id}:
    get:
      tags: [files]
      parameters:
        - $ref: "#/components/parameters/FileID"
      responses:
        "200":
          $ref: "#/components/responses/File"
    delete:
      tags: [files]
      parameters:
        - $ref: "#/components/parameters/FileID"
      responses:
        "200":
          $ref: "#/components/responses/File"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    BranchID:
      name: branch_id
      in: query
      schema:
        type: string
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 100
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    StudentID:
      name: student_id
      in: path
      required: true
      schema:
        type: string
...
    ClassID:
      name: class_id
      in: path
      required: true
      schema:
        type: string
...
    FileID:
      name: file_id
      in: path
      required: true
      schema:
        type: string
...
  requestBodies:
...
    CreateStudent:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [branch_id, fin, first_name, last_name]
            properties:
              branch_id:
                type: string
              fin:
                type: string
                minLength: 7
                maxLength: 7
              first_name:
                type: string
              last_name:
                type: string
              birth_date:
                type: string
                example: "15/07/1998"
              gender:
                enum: [male, female, other]
              phone:
                type: string
              address:
                type: string
              profile_photo_file_id:
                type: string
              status:
                enum: [active, left, graduated]
              parents:
                type: array
                items:
                  type: object
                  required: [relation, name, phones]
                  properties:
                    relation:
                      enum: [father, mother, sister, brother, other]
                    name:
                      type: string
                    phones:
                      type: array
                      items:
                        type: string
              courses:
                type: array
                items:
                  type: object
                  required: [course_id, monthly_amount_cents, start_date]
                  properties:
                    course_id:
                      type: string
                    teacher_id:
                      type: string
                    monthly_amount_cents:
                      type: integer
                    start_date:
                      type: string
                      example: "06/05/2026"
    UpdateStudent:
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              first_name:
                type: string
              last_name:
                type: string
              birth_date:
                type: string
                example: "15/07/1998"
              gender:
                enum: [male, female, other]
              phone:
                type: string
              address:
                type: string
              profile_photo_file_id:
                type: string
              status:
                enum: [active, left, graduated]
              left_reason:
                type: string
...
    EnrollStudent:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [student_id, joined_at]
            properties:
              student_id:
                type: string
              joined_at:
                type: string
                format: date-time
...
    RegisterFile:
      required: true
      content:
        application/json:
          schema:
            type: object
            required: [branch_id, owner_type, owner_id, category, purpose, original_filename, mime_type, original_size_bytes, stored_size_bytes, original_sha256, storage_bucket, storage_key]
            properties:
              branch_id:
                type: string
              owner_type:
                type: string
              owner_id:
                type: string
              category:
                enum: [standard, special]
              policy:
                enum: [standard-ui, standard-downloadable, special]
              purpose:
                enum: [material, payment_receipt, event, passport, diploma, profile_photo, exam_writing]
              original_filename:
                type: string
              mime_type:
                type: string
              original_size_bytes:
                type: integer
              stored_size_bytes:
                type: integer
              original_sha256:
                type: string
              storage_bucket:
                type: string
              storage_key:
                type: string
  responses:
...
    Student:
      description: Student.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Student"
    StudentList:
      description: Student list.
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/Student"
...
    File:
      description: File metadata.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/FileObject"
    FileList:
      description: File list.
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: "#/components/schemas/FileObject"
...
  schemas:
...
    Student:
      type: object
      properties:
        id:
          type: string
        branch_id:
          type: string
        user_id:
          type: string
        fin:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        birth_date:
          type: string
        gender:
          type: string
        phone:
          type: string
        address:
          type: string
        profile_photo_file_id:
          type: string
        status:
          enum: [active, left, graduated]
        left_reason:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    FileObject:
      type: object
      properties:
        id:
          type: string
        branch_id:
          type: string
        uploader_user_id:
          type: string
        owner_type:
          type: string
        owner_id:
          type: string
        category:
          enum: [standard, special]
        policy:
          enum: [standard-ui, standard-downloadable, special]
        purpose:
          enum: [material, payment_receipt, event, passport, diploma, profile_photo, exam_writing]
        original_filename:
          type: string
        mime_type:
          type: string
        original_size_bytes:
          type: integer
        stored_size_bytes:
          type: integer
        original_sha256:
          type: string
        storage_bucket:
          type: string
        storage_key:
          type: string
        retention_until:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
...